+++ R/json.R
@@ -3,7 +3,7 @@
 #' A JSON serializer that only works on a limited types of R data (`NULL`,
 #' lists, arrays, logical/character/numeric/date/time vectors). Other types of
 #' data will be coerced to character. A character string
-#' wrapped in `JS()` is treated as raw JavaScript, so will not be quoted. The
+#' wrapped in `js()` is treated as raw JavaScript, so will not be quoted. The
 #' function `json_vector()` converts an atomic R vector to JSON.
 #'
 #' Both `NULL` and `NA` are converted to `null`. Named lists are converted to
@@ -31,7 +31,7 @@
 #' tojson(matrix(1:12, 3))
 #'
 #' # literal JS code
-#' tojson(list(a = 1:5, b = JS('function() {return true;}')))
+#' tojson(list(a = 1:5, b = js('function() {return true;}')))
 tojson = function(x) {
   if (inherits(x, 'json')) return(x)
   res = structure(.tojson(x), class = 'json')
@@ -69,7 +69,7 @@ tojson = function(x) {
 
 #' @rdname tojson
 #' @export
-JS = function(x) structure(x, class = 'JS_LITERAL')
+js = function(x) structure(x, class = 'JS_LITERAL')
 
 json_atomic = function(x, to_array = NA) {
   use_quote = !(is.numeric(x) || is.logical(x))